converted ipn to middleware
